feat: Init navbar on close icon click and hide initial one
